messaging_app/requirement.txt
appdirs==1.4.4
asgiref==3.8.1
baron==0.10.1
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.2.1
django-cache-url==3.4.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-environs==0.1.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
environs==14.2.0
inflection==0.5.1
marshmallow==4.0.0
packaging==25.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redbaron==0.9.2
rply==0.7.8
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
